requirements.txt
bcrypt==4.0.1
fastapi
uvicorn
sqlalchemy
passlib
pyjwt
python-multipart